+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support Ubuntu-style dracut initrd images.
+- New sos plugin in `warewulf-sos` subpackage.
 
 ### Fixed